+++ packages/flutter_deck/lib/src/renderers/renderers.dart
@@ -1,0 +1,1 @@
+export 'flutter_deck_slide_image_renderer.dart';